+++ comptest/web/forms.py
@@ -1,0 +1,15 @@
+from django import forms
+
+
+class SubmissionForm(forms.Form):
+    """Form create a new submission"""
+
+    name = forms.CharField()
+    description = forms.CharField()
+    gh_repo = forms.URLField()
+
+
+class UploadForm(forms.Form):
+    """Form to upload a version of a submission to be evaluated"""
+
+    file = forms.FileField()
